Change structure add some docstrings
