+++ src/lib.rs
@@ -6,10 +6,8 @@
 //!             are uses substantives (i.e., omit a `get_` prefix) much like the standard library.
 //!             Callback methods have a `on_` prefix
 
-pub mod iterator;
-pub mod kinematics;
-pub mod visitor;
+pub mod tree;
 
-pub use iterator::TreeIteration;
-pub use visitor::visitable::{Accumulable, Visitable, Visiting};
-pub use visitor::visiting::Visitor;
+pub use tree::iterator::DepthFirstIteration;
+pub use tree::visitable::{Accumulable, Visitable};
+pub use tree::visiting::{Visiting, Visitor};
