--- src/visitor/mod.rs
@@ -1,2 +1,0 @@
-pub mod visitable;
-pub mod visiting;
